push

+++ Assets/Scripts/BattleUI/HealthBarUI.cs
@@ -10,14 +10,14 @@ using UnityEngine.UI;
 		public Gradient gradient;
 		public Image fill;
 
-		public void SetMaxHealth(int maxHealth) {
+		public void SetMaxHealth(float maxHealth) {
 
 			slider.maxValue = maxHealth;
 			slider.value = maxHealth;
 			fill.color = gradient.Evaluate(1f);
 		}
 
-		public void SetHealth(int health) {
+		public void SetHealth(float health) {
 
 			slider.value = health;
 			fill.color = gradient.Evaluate(slider.normalizedValue);
